requirements.txt
pydotenv
chromadb
openai
